Fully provision ZNC with ansible [#2]
ssl_keys.yml now lives in the root, for access by the web-server and
irc-bouncer roles. Same for ssl_certificates.yml, which has a
dictionary of domain certificates as well as the certificate chain
(which may not be the same chain forever, but is for now, at least).

Running ZNC as its own user (with a single-domain ssl key!) should
greatly reduce my exposure in the event of security flaws in ZNC.